Fixed sync permissions bug
