+++ app/src/main/java/com/apptec/registrateapp/mainactivity/MainViewModel.java
@@ -103,7 +103,6 @@ public class MainViewModel extends AndroidViewModel {
         isUserLogged = new MutableLiveData<>(true);
 
         // Pull data from the permission catalog
-        permissionPresenter.pullPermissionCatalog();
         permissionPresenter.syncPermissionsWithNetwork();
